test script for running abscibind on test data
